Allow to optionally specify a list of pipelines to check and in.
Fix #53.

Check and out only fetches the specified list of
pipelines passed in the source configuration if present,
fetch all the pipelines otherwise.

The check now also specifies the team a pipeline belongs to.

Tests for check and in commands are modified and expanded
to take into account the new feature.

Signed-off-by: Alessandro Degano <alessandro.degano@pix4d.com>